+++ src/commands/run.rs
@@ -1,13 +1,12 @@
 //! Implementation of command line option for running server
 
-use crate::commands::tracing;
 use crate::influxdb_ioxd::{self, serving_readiness::ServingReadinessState};
 use clap::arg_enum;
-use core::num::NonZeroU16;
 use data_types::server_id::ServerId;
 use std::{net::SocketAddr, net::ToSocketAddrs, path::PathBuf};
 use structopt::StructOpt;
 use thiserror::Error;
+use trogging::cli::TracingConfig;
 
 /// The default bind address for the HTTP API.
 pub const DEFAULT_API_BIND_ADDR: &str = "127.0.0.1:8080";
@@ -43,233 +42,9 @@ Configuration is loaded from the following sources (highest precedence first):
         - pre-configured default values"
 )]
 pub struct Config {
-    /// Logs: filter directive
-    ///
-    /// Configures log severity level filter, by target.
-    ///
-    /// Simplest options: error, warn, info, debug, trace
-    ///
-    /// Levels for different modules can be specified. For example
-    /// `debug,hyper::proto::h1=info` specifies debug logging for all modules
-    /// except for the `hyper::proto::h1' module which will only display info
-    /// level logging.
-    ///
-    /// Extended syntax provided by `tracing-subscriber` includes span/field
-    /// filters. See <https://docs.rs/tracing-subscriber/0.2.17/tracing_subscriber/filter/struct.EnvFilter.html> for more details.
-    #[structopt(long = "--log-filter", env = "LOG_FILTER", default_value = "warn")]
-    pub log_filter: String,
-
-    /// Logs: filter short-hand
-    ///
-    /// Convenient way to set log severity level filter.
-    /// Overrides `--log-filter`.
-    ///
-    /// -v   'info'
-    ///
-    /// -vv  'debug,hyper::proto::h1=info,h2=info'
-    ///
-    /// -vvv 'trace,hyper::proto::h1=info,h2=info'
-    #[structopt(
-        short = "-v",
-        long = "--verbose",
-        multiple = true,
-        takes_value = false,
-        parse(from_occurrences)
-    )]
-    pub log_verbose_count: u8,
-
-    #[rustfmt::skip]
-    /// Logs: message format
-    ///
-    /// Can be one of:
-    ///
-    /// full: human-readable, single line
-    ///
-    ///   Oct 24 12:55:47.815 ERROR shaving_yaks{yaks=3}: fmt::yak_shave: failed to shave yak yak=3 error=missing yak
-    ///   Oct 24 12:55:47.815 TRACE shaving_yaks{yaks=3}: fmt::yak_shave: yaks_shaved=2
-    ///   Oct 24 12:55:47.815  INFO fmt: yak shaving completed all_yaks_shaved=false
-    ///
-    /// pretty: human-readable, multi line
-    ///
-    ///   Oct 24 12:57:29.387 fmt_pretty::yak_shave: failed to shave yak, yak: 3, error: missing yak
-    ///     at examples/examples/fmt/yak_shave.rs:48 on main
-    ///     in fmt_pretty::yak_shave::shaving_yaks with yaks: 3
-    ///
-    ///   Oct 24 12:57:29.387 fmt_pretty::yak_shave: yaks_shaved: 2
-    ///     at examples/examples/fmt/yak_shave.rs:52 on main
-    ///     in fmt_pretty::yak_shave::shaving_yaks with yaks: 3
-    ///
-    ///   Oct 24 12:57:29.387 fmt_pretty: yak shaving completed, all_yaks_shaved: false
-    ///     at examples/examples/fmt-pretty.rs:19 on main
-    ///
-    /// json: machine-parseable
-    ///
-    ///   {"timestamp":"Oct 24 13:00:00.875","level":"ERROR","fields":{"message":"failed to shave yak","yak":3,"error":"missing yak"},"target":"fmt_json::yak_shave","spans":[{"yaks":3,"name":"shaving_yaks"}]}
-    ///   {"timestamp":"Oct 24 13:00:00.875","level":"TRACE","fields":{"yaks_shaved":2},"target":"fmt_json::yak_shave","spans":[{"yaks":3,"name":"shaving_yaks"}]}
-    ///   {"timestamp":"Oct 24 13:00:00.875","level":"INFO","fields":{"message":"yak shaving completed","all_yaks_shaved":false},"target":"fmt_json"}
-    ///
-    /// logfmt: human-readable and machine-parseable
-    ///
-    ///   level=info msg="This is an info message" target="logging" location="logfmt/tests/logging.rs:36" time=1612181556329599000
-    ///   level=debug msg="This is a debug message" target="logging" location="logfmt/tests/logging.rs:37" time=1612181556329618000
-    ///   level=trace msg="This is a trace message" target="logging" location="logfmt/tests/logging.rs:38" time=1612181556329634000
-    #[structopt(long = "--log-format", env = "LOG_FORMAT", default_value = "full", verbatim_doc_comment)]
-    pub log_format: tracing::LogFormat,
-
-    /// Logs: destination
-    ///
-    /// Can be one of: stdout, stderr
-    //
-    // TODO(jacobmarble): consider adding file path, file rotation, syslog, ?
-    #[structopt(
-        long = "--log-destination",
-        env = "LOG_DESTINATION",
-        default_value = "stdout",
-        verbatim_doc_comment
-    )]
-    pub log_destination: tracing::LogDestination,
-
-    /// Tracing: exporter type
-    ///
-    /// Can be one of: none, jaeger, otlp
-    ///
-    /// When enabled, additional flags are considered (see flags related to OTLP
-    /// and Jaeger), and log output is disabled.
-    //
-    // TODO(jacobmarble): allow logs and traces simultaneously
-    #[structopt(
-        long = "--traces-exporter",
-        env = "TRACES_EXPORTER",
-        default_value = "none"
-    )]
-    pub traces_exporter: tracing::TracesExporter,
-
-    /// Tracing: filter directive
-    ///
-    /// Configures traces severity level filter, by target.
-    ///
-    /// Simplest options: error, warn, info, debug, trace
-    ///
-    /// Levels for different modules can be specified. For example
-    /// `debug,hyper::proto::h1=info` specifies debug tracing for all modules
-    /// except for the `hyper::proto::h1` module which will only display info
-    /// level tracing.
-    ///
-    /// Extended syntax provided by `tracing-subscriber` includes span/field
-    /// filters. See <https://docs.rs/tracing-subscriber/0.2.17/tracing_subscriber/filter/struct.EnvFilter.html> for more details.
-    ///
-    /// No filter by default.
-    #[structopt(long = "--traces-filter", env = "TRACES_FILTER")]
-    pub traces_filter: Option<String>,
-
-    /// Tracing: sampler type
-    ///
-    /// Can be one of:
-    /// always_on, always_off, traceidratio,
-    /// parentbased_always_on, parentbased_always_off, parentbased_traceidratio
-    ///
-    /// These alternatives are described in detail at <https://github.com/open-telemetry/opentelemetry-specification/blob/v1.1.0/specification/sdk-environment-variables.md#general-sdk-configuration>.
-    #[structopt(
-        long = "--traces-sampler",
-        env = "TRACES_SAMPLER",
-        default_value = "parentbased_traceidratio"
-    )]
-    pub traces_sampler: tracing::TracesSampler,
-
-    #[rustfmt::skip]
-    /// Tracing: sampler argument
-    ///
-    /// Valid range: [0.0, 1.0].
-    ///
-    /// Only used if `--traces-sampler` is set to
-    /// parentbased_traceidratio (default) or traceidratio.
-    ///
-    /// With sample parentbased_traceidratio, the following rules apply:
-    /// - if parent is sampled, then all of its children are sampled
-    /// - else sample this portion of traces (0.5 = 50%)
-    ///
-    /// More details about this sampling argument at <https://github.com/open-telemetry/opentelemetry-specification/blob/v1.1.0/specification/sdk-environment-variables.md#general-sdk-configuration>.
-    #[structopt(
-        long = "--traces-sampler-arg",
-        env = "TRACES_SAMPLER_ARG",
-        default_value = "1.0",
-        verbatim_doc_comment
-    )]
-    pub traces_sampler_arg: f64,
-
-    /// Tracing: OTLP (eg OpenTelemetry collector) network hostname
-    ///
-    /// Only used if `--traces-exporter` is "otlp".
-    ///
-    /// Protocol is gRPC. HTTP is not supported.
-    #[structopt(
-        long = "--traces-exporter-otlp-host",
-        env = "TRACES_EXPORTER_OTLP_HOST",
-        default_value = "localhost"
-    )]
-    pub traces_exporter_otlp_host: String,
-
-    /// Tracing: OTLP (eg OpenTelemetry collector) network port
-    ///
-    /// Only used if `--traces-exporter` is "otlp".
-    ///
-    /// Protocol is gRPC. HTTP is not supported.
-    #[structopt(
-        long = "--traces-exporter-otlp-port",
-        env = "TRACES_EXPORTER_OTLP_PORT",
-        default_value = "4317"
-    )]
-    pub traces_exporter_otlp_port: NonZeroU16,
-
-    /// Tracing: Jaeger agent network hostname
-    ///
-    /// Protocol is Thrift/Compact over UDP.
-    ///
-    /// Only used if `--traces-exporter` is "jaeger".
-    #[structopt(
-        long = "--traces-exporter-jaeger-agent-host",
-        env = "TRACES_EXPORTER_JAEGER_AGENT_HOST",
-        default_value = "0.0.0.0"
-    )]
-    pub traces_exporter_jaeger_agent_host: String,
-
-    /// Tracing: Jaeger agent network port
-    ///
-    /// Protocol is Thrift/Compact over UDP.
-    ///
-    /// Only used if `--traces-exporter` is "jaeger".
-    #[structopt(
-        long = "--traces-exporter-jaeger-agent-port",
-        env = "TRACES_EXPORTER_JAEGER_AGENT_PORT",
-        default_value = "6831"
-    )]
-    pub traces_exporter_jaeger_agent_port: NonZeroU16,
-
-    /// Tracing: Jaeger service name.
-    ///
-    /// Only used if `--traces-exporter` is "jaeger".
-    #[structopt(
-        long = "--traces-exporter-jaeger-service-name",
-        env = "TRACES_EXPORTER_JAEGER_SERVICE_NAME",
-        default_value = "iox"
-    )]
-    pub traces_exporter_jaeger_service_name: String,
-
-    /// Tracing: Jaeger max UDP packet size
-    ///
-    /// Default to 1300, which is a safe MTU.
-    ///
-    /// You can increase it to 65000 if the target is a jaeger collector
-    /// on localhost. If so, the batching exporter will be enabled for
-    /// extra efficiency. Otherwise an UDP packet will be sent for each exported span.
-    ///
-    /// Only used if `--traces-exporter` is "jaeger".
-    #[structopt(
-        long = "--traces-exporter-jaeger-max-packet-size",
-        env = "TRACES_EXPORTER_JAEGER_MAX_PACKET_SIZE",
-        default_value = "1300"
-    )]
-    pub traces_exporter_jaeger_max_packet_size: usize,
+    // logging and tracing options
+    #[structopt(flatten)]
+    pub(crate) tracing_config: TracingConfig,
 
     /// The identifier for the server.
     ///
